+++ src/Phan/PluginV2.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Phan;
+
+/**
+ * @deprecated - External plugins should migrate to Phan V2 and PluginV3.
+ * @suppress PhanUnreferencedClass this is deprecated and provided to make migrating to PluginV3 easier
+ */
+class PluginV2 extends PluginV3
+{
+}