+++ src/Phan/PluginV2/AutomaticFixCapability.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace Phan\PluginV2;
+
+/**
+ * @deprecated use PluginV3
+ */
+interface AutomaticFixCapability extends \Phan\PluginV3\AutomaticFixCapability
+{
+}