+++ src/Phan/PluginV2/PostAnalyzeNodeCapability.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Phan\PluginV2;
+
+/**
+ * Use PluginV3
+ * @deprecated
+ */
+interface PostAnalyzeNodeCapability extends \Phan\PluginV3\PostAnalyzeNodeCapability
+{
+}