+++ src/Phan/PluginV2/ReturnTypeOverrideCapability.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace Phan\PluginV2;
+
+/**
+ * @deprecated use PluginV3
+ */
+interface ReturnTypeOverrideCapability extends \Phan\PluginV3\ReturnTypeOverrideCapability
+{
+}