+++ src/Phan/PluginV2/StopParamAnalysisException.php
@@ -1,0 +1,13 @@
+<?php declare(strict_types=1);
+
+namespace Phan\PluginV2;
+
+/**
+ * Use PluginV3
+ * @deprecated use PluginV3
+ * @see \Phan\PluginV3\StopParamAnalysisException
+ * @suppress PhanUnreferencedClass
+ */
+class StopParamAnalysisException extends \Phan\PluginV3\StopParamAnalysisException
+{
+}