+++ src/Phan/PluginV3/AfterAnalyzeFileCapability.php
@@ -10,7 +10,7 @@ use Phan\Language\Context;
  * AfterAnalyzeFileCapability is used when you want to perform checks after analyzing a file
  * NOTE: This does not run on empty files.
  */
-interface AfterAnalyzeFileCapability
+interface AfterAnalyzeFileCapability extends \Phan\PluginV2\AfterAnalyzeFileCapability
 {
     /**
      * This method is called after Phan analyzes a file.