+++ src/Phan/PluginV3/AnalyzeMethodCapability.php
@@ -9,7 +9,7 @@ use Phan\Language\Element\Method;
  * Plugins can implement this to analyze (and modify) a method definition,
  * after parsing and before analyzing.
  */
-interface AnalyzeMethodCapability
+interface AnalyzeMethodCapability extends \Phan\PluginV2\AnalyzeMethodCapability
 {
     /**
      * Analyze (and modify) a method definition, after parsing and before analyzing.